pyproject.toml
[project]
name = "living-memory"
version = "0.1.0"
description = "A static website generator managed by AI agent memory"
requires-python = ">=3.11"
dependencies = [
    "python-frontmatter>=1.1.0",
    "google-genai>=1.0.0",
]

[project.scripts]
publisher = "publisher:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
